+++ src/Mod/Part/Gui/TaskOffset.cpp
@@ -45,7 +45,7 @@
 #include <App/Application.h>
 #include <App/Document.h>
 #include <App/DocumentObject.h>
-#include <Mod/Part/App/PartFeatures.h>
+#include <Mod/Part/App/FeatureOffset.h>
 
 
 using namespace PartGui;
